Split out DataSerializable interface.
