+++ tests/cpp/testobject/DeltaTestImpl.hpp
@@ -26,9 +26,11 @@
 #include <ace/Condition_T.h>
 #include <ace/Recursive_Thread_Mutex.h>
 #include <ace/Time_Value.h>
-#include "TestObject1.hpp"
-#include <geode/Delta.hpp>
 
+#include <geode/Delta.hpp>
+#include <geode/DataSerializable.hpp>
+
+#include "TestObject1.hpp"
 #include "testobject_export.h"
 
 using namespace apache::geode::client;
