+++ tests/cpp/testobject/Position.hpp
@@ -27,6 +27,7 @@
 #include <string>
 
 #include <geode/CacheableString.hpp>
+#include <geode/DataSerializable.hpp>
 
 #include "testobject_export.h"
