+++ tests/cpp/testobject/VariousPdxTypes.hpp
@@ -20,6 +20,7 @@
 #ifndef GEODE_TESTOBJECT_VARIOUSPDXTYPES_H_
 #define GEODE_TESTOBJECT_VARIOUSPDXTYPES_H_
 
+#include <geode/DataSerializable.hpp>
 #include <geode/PdxSerializable.hpp>
 #include <geode/PdxWriter.hpp>
 #include <geode/PdxReader.hpp>
